Added Authentication

+++ src/AuctionService/Controllers/AuctionsEndpoints.cs
@@ -18,32 +18,28 @@ public class AuctionsEndpoints : ICarterModule
 
         group.MapGet("", GetAuctions).WithName("GetAuctions");
         group.MapGet("{id:guid}", GetAuction).WithName("GetAuction");
-        group.MapPost("", CreateAuction);
-        group.MapPut("{id:guid}", UpdateAuction);
-        group.MapDelete("{id:guid}", DeleteAuction);
+        group.MapPost("", CreateAuction).RequireAuthorization();
+        group.MapPut("{id:guid}", UpdateAuction).RequireAuthorization();
+        group.MapDelete("{id:guid}", DeleteAuction).RequireAuthorization();
     }
 
     private static async Task<IResult> CreateAuction(
         CreateAuctionDto request, 
         DataContext context,
         IMapper mapper,
+        HttpContext httpContext,
         IPublishEndpoint publishEndpoint)
     {
         var auction = mapper.Map<Auction>(request);
-        // userId
-        auction.Seller = "test";
-        
+        auction.Seller = httpContext.User.Identity.Name;
         context.Auctions.Add(auction);
         
         var newAuction = mapper.Map<AuctionDto>(auction);
-        
         await publishEndpoint.Publish(mapper.Map<AuctionCreated>(newAuction));
-        
         var result = await context.SaveChangesAsync() > 0;
-
-
-        return result ?
-            Results.CreatedAtRoute(nameof(GetAuction), new { auction.Id }, newAuction) :
+        
+        return result ? 
+            Results.CreatedAtRoute(nameof(GetAuction), new { auction.Id }, newAuction) : 
             Results.BadRequest("Could not save changes to DB");
     }
 
@@ -75,42 +71,43 @@ public class AuctionsEndpoints : ICarterModule
     private static async Task<IResult> UpdateAuction(
         UpdateAuctionDto request,
         DataContext context,
-        IMapper mapper, Guid id,
+        IMapper mapper,
+        Guid id,
+        HttpContext httpContext,
         IPublishEndpoint publishEndpoint)
     {
         var auction = await context.Auctions.Include(x => x.Item)
             .FirstOrDefaultAsync(x => x.Id == id);
         
         if(auction is null) return Results.NotFound();
+        if(auction.Seller != httpContext.User.Identity.Name) return Results.Forbid();
         
-        //userId
         auction.Item.Make = request.Make ?? auction.Item.Make;
         auction.Item.Model = request.Model ?? auction.Item.Model;
         auction.Item.Color= request.Color ?? auction.Item.Color;
         auction.Item.Mileage = request.Mileage ?? auction.Item.Mileage;
         auction.Item.Year = request.Year ?? auction.Item.Year;
-        //auction = mapper.Map<Auction>(request);
         
         await publishEndpoint.Publish(mapper.Map<AuctionUpdated>(auction));
-        
         var result = await context.SaveChangesAsync() > 0;
         
         return result ? Results.Ok() : Results.BadRequest("Could not save changes to DB");
     }
 
     private static async Task<IResult> DeleteAuction(
-        DataContext context, Guid id,
+        DataContext context,
+        HttpContext httpContext,
+        Guid id,
         IPublishEndpoint publishEndpoint)
     {
         var auction = await context.Auctions.FindAsync(id);
         
         if(auction is null) return Results.NotFound();
+        if(auction.Seller != httpContext.User.Identity.Name) return Results.Forbid();
         
-        //userId
         context.Auctions.Remove(auction);
         
         await publishEndpoint.Publish(new AuctionDeleted { Id = auction.Id });
-        
         var result = await context.SaveChangesAsync() > 0;
         
         return result ? Results.Ok() : Results.BadRequest("Could not save changes to DB");
